requirements.txt
nltk>=3.8
sentence-transformers>=2.6
hdbscan>=0.8
click>=8.0
markdown-it-py>=2.2
pyyaml>=6.0
jsonschema>=4.21
rich>=13.7
pytest
scikit-learn>=1.4
torch>=2.2  